# Input/output permutation for fused attention