Depends on #436 #429

This PR aims to update `BatchedGemmSoftmaxGemmPermute` and `GroupedGemmSoftmaxGemmPermute` in line with [device_batched_contraction_multiple_d_xdl_cshuffle.hpp](https://github.com/ROCmSoftwarePlatform/composable_kernel/blob/b8825547586855ec730a2eca47e415b1404bb5f2/include/ck/tensor_operation/gpu/device/device_batched_contraction_multiple_d_xdl_cshuffle.hpp). The old arguments will be replaced by vector of lengths and strides of each tensor, and provides "Packed" tensor specialization for degenerating each N-D tensor into 3D form [G, M, N] when permutation isn't needed.

"Packed" in a sense that each inner dimension in a dimension group (eg [G0, G1] in Gs, [M0, M1, M2] in Ms, etc.) are contiguous and ordered. Not in a sense that the tensor [G0, G1, ..., M0, M1, ..., N0, N1...] can be permuted while still being a contiguous, unpadded tensor.

- [x] BatchedGemmSoftmaxGemmPermute example code working
- [x] BatchedGemmSoftmaxGemmPermute test code working
- [x] Tensor specialization to avoid paying extra cost when permutation isn't needed
- [x] GroupedGemmSoftmaxGemmPermute example code working
- [x] Test if any perf regression

Run with `$ build/bin/test_batched_gemm_softmax_gemm_permute_fp16 --gtest_filter=*Bench_FP16 --gtest_also_run_disabled_tests | grep 'c_gs_ms_os|Best Perf' -P`:
| batch_size | nheads | batch_size*nheads | seqlen | head_dim | ms (new) | ms (old) | Diff |
|------------|--------|-------------------|--------|----------|----------|----------|------|
| 48         | 16     | 768               | 256    | 64       | 0.212273 | 0.199505 | 6%   |
| 48         | 16     | 768               | 256    | 128      | 0.465234 | 0.478194 | -3%  |
| 48         | 16     | 768               | 512    | 64       | 0.645763 | 0.626611 | 3%   |
| 48         | 16     | 768               | 512    | 128      | 1.42924  | 1.47793  | -3%  |
| 48         | 16     | 768               | 1024   | 64       | 2.29079  | 2.23515  | 2%   |
| 48         | 16     | 768               | 1024   | 128      | 4.81568  | 4.95164  | -3%  |
| 48         | 16     | 768               | 2048   | 64       | 8.80842  | 8.66246  | 2%   |
| 48         | 16     | 768               | 2048   | 128      | 17.941   | 18.3959  | -2%  |
| 48         | 16     | 768               | 4096   | 64       | 34.2197  | 34.3561  | 0%   |
| 48         | 16     | 768               | 4096   | 128      | 71.0401  | 72.4781  | -2%  |

Next actions:
- Acc0Bias, Acc1Bias